Implement Mesh Subdivision Algorithm
In order to generate our asteroid, I realized I can
programmatically create a tetrahedron. Then I can
iteratively subdivide it then add noise to it. In
this commit, I implemented an algorithm to subdivide
a triangle mesh using the Loop Subdivision Algorithm.

https://graphics.stanford.edu/~mdfisher/subdivision.html

Next step is to incorporate this subdivision algorithm in
the asteroid generator.