Add Material property to Models
Model's textures and shaders were separate before, I've
combined them, as well as a few new properties such as
object color and phong highlight properties to make a new
class called Material. This encapsulates a lot of logic that
was floating around everywhere. In doing so, I've made a class
called ShaderProgram to represent a shader and some of its
common operations instead of having a plain int floating around.