Begin separating Rendering from GameEngine
We want to possibly create a 3D renderer. So I am separating
the rendering from the GameEngine itself. We can then
substitute a 3D renderer for a normal renderer.

At first I thought we'd want a Camera and a Camera3D, but
rendering in 3D is a lot more than just changing a camera,
so I figured it warranted a separate class for how you render.
This could also be extended somehow to apply post-processing
effects I imagine.