Complete Renderer3D and Fix Multiple Shaders
Renderer3D works now! Just change the renderer in the GameEngine.
Options are Renderer, FrameBufferRenderer (no noticeable difference),
and Renderer3D.

We can extend FrameBufferRenderer and Renderer3D to perform
postprocessing effects such as bloom if we really want to. Oooo!
Cool idea: we could make the screen flash red or something when
we get hit!

To view the game in 3D, I put my phone into a google cardboard device
and use Chrome Remote Desktop to mirror my computer to my phone.
It is legit!

There was a bug where multiple shaders didn't work. That's because
I was trying to set the uniforms in all the shaders and then tried
to render all the objects. Obviously that doesn't work because as
you set stuff for one shader, you erase stuff from a previous shader.
I made sort of a hack that for each shader it sets all the uniforms
then traverses all the objects and only draws those that use that
shader.
It requires multiple traversals of our object tree, but I didn't
see any noticeable slow down. Maybe it would cause that with
millions of objects...