Adjust VR Camera centering
For virtual reality, the left camera was in the center, and the right
camera was off to the right. I adjusted it so that the left camera
is off to the left a bit and the right camera is off to to right a bit.
This will give us a more correct 3d experience.